requirements.txt
fastapi[all]
uvicorn
pymongo
python-dotenv
passlib[bcrypt]
pydantic
python-jose
PyJWT
